+++ tests/testthat/test-conversions.R
@@ -29,7 +29,7 @@ test_that("from STM", {
     apply(stm$theta, 2, function(x) head(order(-x)))
   )
 
-  expect_error(grow(lda), "Not possible for approximated")
+  expect_error(fit(lda), "Not possible for approximated")
   expect_error(mergeTopics(lda, as.list(1:3)), "Not possible for approximated")
   expect_error(rJST(lda), "Not possible for approximated")
 
@@ -177,12 +177,12 @@ test_that("from keyATM", {
   skip_if_not_installed("keyATM")
   library("keyATM")
   library("quanteda")
-  
+
   data(keyATM_data_bills)
   bills_keywords <- keyATM_data_bills$keywords
   bills_dfm <- keyATM_data_bills$doc_dfm  # quanteda dfm object
   keyATM_docs <- keyATM_read(bills_dfm)
-  
+
   # keyATM Base
   out <- keyATM(
     docs = keyATM_docs,
@@ -192,17 +192,17 @@ test_that("from keyATM", {
     options = list(iterations = 100)
   )
   LDA <- as.LDA(out, keyATM_docs)
-  
+
   expect_true(check_integrity(LDA))
   expect_silent(topWords(LDA))
-  
+
   # check top words
   out$phi <- out$phi[, order(colnames(out$phi))] # force correct order
   expect_identical(
     topWords(LDA, output = "matrix", method = "probability"),
     as.matrix(top_words(out, show_keyword = FALSE))
   )
-  
+
   # check top documents
   expect_equal(
     apply(LDA$theta, 2, function(x) head(order(-x))),
@@ -216,7 +216,7 @@ test_that("LDAvis", {
   skip_if_not_installed("servr")
 
   lda <- LDA(ECB_press_conferences_tokens)
-  lda <- grow(lda, 10, displayProgress = FALSE)
+  lda <- fit(lda, 10, displayProgress = FALSE)
   if (interactive())
     expect_message(LDAvis(lda), "To stop the server,")
   else