+++ tests/testthat/test-generics.R
@@ -1,5 +1,5 @@
 
-context("Tests for coherence metrics")
+context("Tests generics")
 toks <- ECB_press_conferences_tokens[quanteda::ntoken(ECB_press_conferences_tokens) >= 110]
 dfm <- quanteda::dfm(toks, tolower = FALSE)
 dfm <- quanteda::dfm_trim(dfm, min_termfreq = 2)