+++ NEWS.md
@@ -1,3 +1,10 @@
+# sentopics 0.7.3
+
+* Renamed `grow()` to `fit()` as a more intuitive name. `grow()` remains in the package for compatibility with older version.
+* Aligned the weight argument and the algorithm of `topWords(..., method="FREX")` to the original paper.
+* Fixed the un-exported function `get_ECB_conferences()` to accommodates changes from the ECB website.
+* Fixed the model conversion from newer `seededlda` versions (1.2.0 and subsequent)
+
 # sentopics 0.7.2
 
 * Added `as.LDA()` method for outputs from the **keyATM** package.